requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pydantic
python-dotenv
psycopg2-binary
python-multipart
email-validator
jinja2
authlib
httpx
itsdangerous
